Fix AuthBasic browser error exports
